+++ NEWS.md
@@ -3,7 +3,8 @@
 
 * `rim rm now supports removing multiple versions.
 
-* `rim system create-lib` can be now restricted to one or more R versions.
+* `rim system` commands `create-lib` and `fix-permissions` can be now
+  restricted to one or more R versions.
 
 # rim 0.1.0